# openapi.yaml
openapi: 3.0.3
info:
  title: SoftDesk API
  version: 1.0.0
  description: API REST pour gérer des projets, issues et commentaires.
paths:
  /api/login/:
    post:
      operationId: login_create
      description: |-
        Takes a set of user credentials and returns an access and refresh JSON web
        token pair to prove the authentication of those credentials.
      tags:
      - login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenObtainPair'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TokenObtainPair'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TokenObtainPair'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenObtainPair'
          description: ''
  /api/projects/:
    get:
      operationId: projects_list
      tags:
      - projects
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
          description: ''
    post:
      operationId: projects_create
      tags:
      - projects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Project'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Project'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: ''
  /api/projects/{id}/:
    get:
      operationId: projects_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - projects
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: ''
    put:
      operationId: projects_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - projects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Project'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Project'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: ''
    patch:
      operationId: projects_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - projects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedProject'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedProject'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedProject'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: ''
    delete:
      operationId: projects_destroy
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - projects
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
  /api/projects/{project_pk}/issues/:
    get:
      operationId: projects_issues_list
      parameters:
      - in: path
        name: project_pk
        schema:
          type: integer
        required: true
      tags:
      - projects
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Issue'
          description: ''
    post:
      operationId: projects_issues_create
      parameters:
      - in: path
        name: project_pk
        schema:
          type: integer
        required: true
      tags:
      - projects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Issue'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Issue'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Issue'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
          description: ''
  /api/projects/{project_pk}/issues/{issue_pk}/comments/:
    get:
      operationId: projects_issues_comments_list
      parameters:
      - in: path
        name: issue_pk
        schema:
          type: integer
        required: true
      - in: path
        name: project_pk
        schema:
          type: integer
        required: true
      tags:
      - projects
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
          description: ''
    post:
      operationId: projects_issues_comments_create
      parameters:
      - in: path
        name: issue_pk
        schema:
          type: integer
        required: true
      - in: path
        name: project_pk
        schema:
          type: integer
        required: true
      tags:
      - projects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comment'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Comment'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Comment'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
          description: ''
  /api/projects/{project_pk}/issues/{issue_pk}/comments/{id}/:
    get:
      operationId: projects_issues_comments_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      - in: path
        name: issue_pk
        schema:
          type: integer
        required: true
      - in: path
        name: project_pk
        schema:
          type: integer
        required: true
      tags:
      - projects
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
          description: ''
    put:
      operationId: projects_issues_comments_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      - in: path
        name: issue_pk
        schema:
          type: integer
        required: true
      - in: path
        name: project_pk
        schema:
          type: integer
        required: true
      tags:
      - projects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comment'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Comment'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Comment'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
          description: ''
    patch:
      operationId: projects_issues_comments_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      - in: path
        name: issue_pk
        schema:
          type: integer
        required: true
      - in: path
        name: project_pk
        schema:
          type: integer
        required: true
      tags:
      - projects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedComment'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedComment'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedComment'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
          description: ''
    delete:
      operationId: projects_issues_comments_destroy
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      - in: path
        name: issue_pk
        schema:
          type: integer
        required: true
      - in: path
        name: project_pk
        schema:
          type: integer
        required: true
      tags:
      - projects
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
  /api/projects/{project_pk}/issues/{id}/:
    get:
      operationId: projects_issues_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      - in: path
        name: project_pk
        schema:
          type: integer
        required: true
      tags:
      - projects
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
          description: ''
    put:
      operationId: projects_issues_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      - in: path
        name: project_pk
        schema:
          type: integer
        required: true
      tags:
      - projects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Issue'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Issue'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Issue'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
          description: ''
    patch:
      operationId: projects_issues_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      - in: path
        name: project_pk
        schema:
          type: integer
        required: true
      tags:
      - projects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedIssue'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedIssue'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedIssue'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
          description: ''
    delete:
      operationId: projects_issues_destroy
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      - in: path
        name: project_pk
        schema:
          type: integer
        required: true
      tags:
      - projects
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
  /api/projects/{project_pk}/users/:
    get:
      operationId: projects_users_list
      description: |-
        Manage project collaborators.

        Routes:
        - GET /projects/{project_pk}/users/
        - POST /projects/{project_pk}/users/
        - DELETE /projects/{project_pk}/users/{user_pk}/
      parameters:
      - in: path
        name: project_pk
        schema:
          type: integer
        required: true
      tags:
      - projects
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contributor'
          description: ''
    post:
      operationId: projects_users_create
      description: |-
        Manage project collaborators.

        Routes:
        - GET /projects/{project_pk}/users/
        - POST /projects/{project_pk}/users/
        - DELETE /projects/{project_pk}/users/{user_pk}/
      parameters:
      - in: path
        name: project_pk
        schema:
          type: integer
        required: true
      tags:
      - projects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Contributor'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Contributor'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Contributor'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contributor'
          description: ''
  /api/projects/{project_pk}/users/{user_pk}/:
    delete:
      operationId: projects_users_destroy
      description: |-
        Manage project collaborators.

        Routes:
        - GET /projects/{project_pk}/users/
        - POST /projects/{project_pk}/users/
        - DELETE /projects/{project_pk}/users/{user_pk}/
      parameters:
      - in: path
        name: project_pk
        schema:
          type: integer
        required: true
      - in: path
        name: user_pk
        schema:
          type: integer
        required: true
      tags:
      - projects
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
  /api/signup/:
    post:
      operationId: signup_create
      tags:
      - signup
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Signup'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Signup'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Signup'
        required: true
      security:
      - jwtAuth: []
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Signup'
          description: ''
  /api/token/:
    post:
      operationId: token_create
      description: |-
        Takes a set of user credentials and returns an access and refresh JSON web
        token pair to prove the authentication of those credentials.
      tags:
      - token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenObtainPair'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TokenObtainPair'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TokenObtainPair'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenObtainPair'
          description: ''
  /api/token/refresh/:
    post:
      operationId: token_refresh_create
      description: |-
        Takes a refresh type JSON web token and returns an access type JSON web
        token if the refresh token is valid.
      tags:
      - token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefresh'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TokenRefresh'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TokenRefresh'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenRefresh'
          description: ''
components:
  schemas:
    Comment:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        description:
          type: string
        author_user:
          allOf:
          - $ref: '#/components/schemas/UserSummary'
          readOnly: true
        issue:
          type: integer
          readOnly: true
        created_time:
          type: string
          format: date-time
          readOnly: true
      required:
      - author_user
      - created_time
      - description
      - id
      - issue
    Contributor:
      type: object
      description: |-
        Serialize project contributors.

        The API only accepts a user identifier when adding a contributor. The
        project, role and permission values are server-controlled, which prevents a
        caller from elevating privileges through mass assignment.
      properties:
        id:
          type: integer
          readOnly: true
        user:
          allOf:
          - $ref: '#/components/schemas/UserSummary'
          readOnly: true
        user_id:
          type: integer
          writeOnly: true
        project:
          type: integer
          readOnly: true
        permission:
          allOf:
          - $ref: '#/components/schemas/PermissionEnum'
          readOnly: true
        role:
          allOf:
          - $ref: '#/components/schemas/RoleEnum'
          readOnly: true
        created_time:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_time
      - id
      - permission
      - project
      - role
      - user
      - user_id
    Issue:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        title:
          type: string
          maxLength: 255
        description:
          type: string
        author_user:
          allOf:
          - $ref: '#/components/schemas/UserSummary'
          readOnly: true
        assignee_user:
          type: integer
          nullable: true
        project:
          type: integer
          readOnly: true
        priority:
          $ref: '#/components/schemas/PriorityEnum'
        tag:
          $ref: '#/components/schemas/TagEnum'
        status:
          $ref: '#/components/schemas/StatusEnum'
        created_time:
          type: string
          format: date-time
          readOnly: true
      required:
      - author_user
      - created_time
      - description
      - id
      - project
      - title
    PatchedComment:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        description:
          type: string
        author_user:
          allOf:
          - $ref: '#/components/schemas/UserSummary'
          readOnly: true
        issue:
          type: integer
          readOnly: true
        created_time:
          type: string
          format: date-time
          readOnly: true
    PatchedIssue:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        title:
          type: string
          maxLength: 255
        description:
          type: string
        author_user:
          allOf:
          - $ref: '#/components/schemas/UserSummary'
          readOnly: true
        assignee_user:
          type: integer
          nullable: true
        project:
          type: integer
          readOnly: true
        priority:
          $ref: '#/components/schemas/PriorityEnum'
        tag:
          $ref: '#/components/schemas/TagEnum'
        status:
          $ref: '#/components/schemas/StatusEnum'
        created_time:
          type: string
          format: date-time
          readOnly: true
    PatchedProject:
      type: object
      description: |-
        Serialize projects without allowing client-side ownership changes.

        The project author is always set from the authenticated request user in
        ProjectViewSet.perform_create. Contributors are managed only through the
        dedicated contributor endpoints to avoid mass-assignment bypasses.
      properties:
        id:
          type: integer
          readOnly: true
        title:
          type: string
          maxLength: 255
        description:
          type: string
        type:
          $ref: '#/components/schemas/TypeEnum'
        author_user:
          allOf:
          - $ref: '#/components/schemas/UserSummary'
          readOnly: true
        contributors:
          type: array
          items:
            $ref: '#/components/schemas/UserSummary'
          readOnly: true
        created_time:
          type: string
          format: date-time
          readOnly: true
    PermissionEnum:
      enum:
      - READ
      - WRITE
      - ADMIN
      type: string
      description: |-
        * `READ` - Read
        * `WRITE` - Write
        * `ADMIN` - Admin
    PriorityEnum:
      enum:
      - LOW
      - MEDIUM
      - HIGH
      type: string
      description: |-
        * `LOW` - Low
        * `MEDIUM` - Medium
        * `HIGH` - High
    Project:
      type: object
      description: |-
        Serialize projects without allowing client-side ownership changes.

        The project author is always set from the authenticated request user in
        ProjectViewSet.perform_create. Contributors are managed only through the
        dedicated contributor endpoints to avoid mass-assignment bypasses.
      properties:
        id:
          type: integer
          readOnly: true
        title:
          type: string
          maxLength: 255
        description:
          type: string
        type:
          $ref: '#/components/schemas/TypeEnum'
        author_user:
          allOf:
          - $ref: '#/components/schemas/UserSummary'
          readOnly: true
        contributors:
          type: array
          items:
            $ref: '#/components/schemas/UserSummary'
          readOnly: true
        created_time:
          type: string
          format: date-time
          readOnly: true
      required:
      - author_user
      - contributors
      - created_time
      - description
      - id
      - title
      - type
    RoleEnum:
      enum:
      - AUTHOR
      - CONTRIBUTOR
      type: string
      description: |-
        * `AUTHOR` - Author
        * `CONTRIBUTOR` - Contributor
    Signup:
      type: object
      description: Create a user account using Django password validators.
      properties:
        id:
          type: integer
          readOnly: true
        username:
          type: string
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
          pattern: ^[\w.@+-]+$
          maxLength: 150
        email:
          type: string
          format: email
        password:
          type: string
          writeOnly: true
          minLength: 8
      required:
      - email
      - id
      - password
      - username
    StatusEnum:
      enum:
      - TODO
      - IN_PROGRESS
      - COMPLETED
      type: string
      description: |-
        * `TODO` - To-Do
        * `IN_PROGRESS` - In Progress
        * `COMPLETED` - Completed
    TagEnum:
      enum:
      - BUG
      - ENHANCEMENT
      - TASK
      type: string
      description: |-
        * `BUG` - Bug
        * `ENHANCEMENT` - Enhancement
        * `TASK` - Task
    TokenObtainPair:
      type: object
      properties:
        username:
          type: string
          writeOnly: true
        password:
          type: string
          writeOnly: true
        access:
          type: string
          readOnly: true
        refresh:
          type: string
          readOnly: true
      required:
      - access
      - password
      - refresh
      - username
    TokenRefresh:
      type: object
      properties:
        access:
          type: string
          readOnly: true
        refresh:
          type: string
          writeOnly: true
      required:
      - access
      - refresh
    TypeEnum:
      enum:
      - BACKEND
      - FRONTEND
      - IOS
      - ANDROID
      type: string
      description: |-
        * `BACKEND` - Back-end
        * `FRONTEND` - Front-end
        * `IOS` - iOS
        * `ANDROID` - Android
    UserSummary:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        username:
          type: string
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
          pattern: ^[\w.@+-]+$
          maxLength: 150
      required:
      - id
      - username
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT